Add block fall speed to the config Make pieces fall and when stops moving loads the next piece Add area to the Piece class to facilitate maths

+++ BlockFall/source/Game/GameField.h
@@ -6,5 +6,5 @@ struct GameField
 {
     float previewWindowSize = 100.0f;
     std::unique_ptr<Board> board = std::make_unique<Board>();
-    SDL_FPoint previewZoneOrigin = { Config::getInstance().getConfigData().width - 200.0f, Config::getInstance().getConfigData().height - 200.0f};
+    SDL_FPoint previewZoneOrigin = { static_cast<float>(Config::getInstance().getConfigData().width) - 200.0f, static_cast<float>(Config::getInstance().getConfigData().height) - 200.0f};
 };
